+++ packages/components/src/components/search/search.tsx
@@ -35,6 +35,11 @@ export class Search implements ComponentInterface {
     @Event() shortInput: EventEmitter<void>;
 
     /**
+     * Event emitted whenever the value of the input field has changed.
+     */
+    @Event() changed: EventEmitter<void>;
+
+    /**
      * Placeholder for the input field.
      */
     @Prop() placeholder: string = '';
@@ -254,6 +259,7 @@ export class Search implements ComponentInterface {
             if (inputValue.length > 1 && inputValue !== this.lastRequested) {
                 this.search(inputValue);
             }
+            this.changed.emit();
         }
     }