+++ packages/map-template/src/components/WebComponentWrappers/Search/Search.jsx
@@ -13,12 +13,13 @@ import sessionTokenState from '../../../atoms/sessionTokenState';
  * @param {function} props.results - Function that is called when search results are received.
  * @param {function} props.clicked - Function that is called when search field is clicked.
  * @param {function} props.cleared - Function that is called when search field is cleared.
+ * @param {function} props.changed - Function that is called when value of the input field is changed.
  * @param {string} props.category - If set, search will be performed for Locations having this category.
  * @param {boolean} props.preventFocus - If set to true, the search field will be disabled.
  * @param {boolean} props.google - Set to true to include results from Google Places autocomplete service.
  * @param {boolean} props.mapbox - Set to true to include results from Mapbox Places autocomplete service.
  */
-const SearchField = forwardRef(({ placeholder, mapsindoors, results, clicked, cleared, category, google, mapbox, disabled = false }, ref) => {
+const SearchField = forwardRef(({ placeholder, mapsindoors, results, clicked, cleared, changed, category, google, mapbox, disabled = false }, ref) => {
     const elementRef = useRef();
 
     const mapboxAccessToken = useRecoilValue(mapboxAccessTokenState);
@@ -67,14 +68,16 @@ const SearchField = forwardRef(({ placeholder, mapsindoors, results, clicked, cl
         current.addEventListener('results', searchResultsHandler);
         current.addEventListener('click', clicked);
         current.addEventListener('cleared', onCleared);
+        current.addEventListener('changed', changed);
 
         return () => {
             current.removeEventListener('results', searchResultsHandler);
             current.removeEventListener('click', clicked);
             current.removeEventListener('cleared', onCleared);
+            current.removeEventListener('changed', changed);
         }
 
-    }, [placeholder, mapsindoors, results, clicked, cleared, google, mapbox]);
+    }, [placeholder, mapsindoors, results, clicked, cleared, google, mapbox, changed]);
 
     return <mi-search ref={elementRef}
         placeholder={placeholder}